+++ product-service/src/main/java/com/ttdat/productservice/domain/events/product/ProductUpdatedEvent.java
@@ -19,7 +19,7 @@ public class ProductUpdatedEvent {
 
     String description;
 
-    Integer totalQuantity;
+    Double totalQuantity;
 
     String imageUrl;
